refactor: extract calendar and rename tenure card
