+++ components/dashboard-card/time-as-engineer.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 
-import { DaysAsSoftwareEngineer } from '@/components/dashboard-card/days-as-engineer';
+import { TimeAsSoftwareEngineer } from '@/components/dashboard-card/time-as-engineer';
 
-describe('DaysAsSoftwareEngineer', () => {
+describe('TimeAsSoftwareEngineer', () => {
   beforeEach(() => {
     jest.useFakeTimers().setSystemTime(new Date('2026-07-19T12:00:00Z'));
   });
@@ -12,7 +12,7 @@ describe('DaysAsSoftwareEngineer', () => {
   });
 
   it('should render elapsed time as years, months and days', () => {
-    render(<DaysAsSoftwareEngineer />);
+    render(<TimeAsSoftwareEngineer />);
 
     expect(screen.getByText('7y, 11m, 18d')).toBeInTheDocument();
   });
